New HeterogeneousExpandableList interface.
This interface can be used in conjunction with ExpandableListAdapter to
declare more types of child and/or group views.

None of the ExpandableListAdapter implementations is heterogeneous in the
framework. BaseExpandableListAdapter was decalred to use this interface so that users
see the methods and can overload. DateSortedExpandableListAdapter was left unchanged.

This feature is related to http://b/issue?id=1459940

Change-Id: Ifc589b697913778b16abfdcaaa9f8f81e564add7