Fix problem with calling onConfigurationChanged() too much.
We now return the initial configuration for a window when it is
added to the window manager.  The view hierarchy would check to
see if it was different than the last one, and not dispatch a
configuration change down itself if not.  However, when
ActivityThread received it, it would always dispatch a config
change even if it is the same.

The solution is to only do this in ActivityThread if the config
is actually different; otherwise, we continue to rely only on
the activity manager explicitly telling us when to do a config
change.

Change-Id: I8a6e3565776dd2723c8b791496bb6041463d4b67