Log full exception when failing to inflate notification view
When inflating a notification's view fails, include the exception in
the log message.  Without this exception all we get is "couldn't
inflate view for notification <package>/<id>", which isn't very
helpful for tracking down the particular error in the view.

This exception used to be included in the log message, but it was
removed in 8a6361e8597 -- any particular reason why?

Change-Id: I623b9e4c8291e4c035f26380e5f22ad6b65176a7