Move the "voice_capable" resource into the framework.
There's currently a config.xml param in the phone app saying whether the
current device is "voice capable", which is used to determine (among other
things) whether the OTASP call should run in interactive mode.

But both Contacts and Settings are about to need this flag too, so it
really belongs in the framework instead.  As of this change it's now
"com.android.internal.R.bool.voice_capable".

Also add a corresponding TelephonyManager call so that unbundled and 3rd
party apps can access it too.

Change-Id: Ie9a8803295bf97eb32154291aac8fce1e21c7b25