simplify cursor ring draw state; fix null check
Rename jni call to note that, when pressing the dpad, the
cursor ring is to shown for a short time. The old name suggested
that this was tracking whether the link was followed.

Fix an unrelated change (from gingerbread, automated
merge unlikely to succeed) to check for a null
WebChromeClient.

Requires a companion change in external/webkit

http://b/2135321 (rename jni call)
http://b/2945171 (null check)

Change-Id: I9d4a92dbba1f92da8dc8cde755a2066de59514f9