Add LocalePicker fragment as one of internal components.
LocalePicker is in Settings, but we want this in framework
so that we are able to use it from other Setting-like
applications.

Also brings two string-arrays for descripting Chinise locales, which
have been in Settings app.

See also I1364c4810a115958662ca5d6e6e7f29c0c9b2569, for example.

Change-Id: I8a398b9b77601cf88ea8a3840b92dcf7441e6103