Fixes a problem with URL parsing when the host ends with a dot
The WebAddress class provides a lenient parser for URLs. Currently, it
identifies the host portion with the regex [<chars>]+(\.[<chars>]+)* where
<chars> is the set of characters valid for the host name. This pattern excludes
the case where the host ends with a dot, which is valid possibility. As a
result, any trailing dot is pushed into the path component. Since we add a
leading slash to the path if one is missing, the result is a path that begins
with '/.'.

This fix changes the host regex to [<chars>]+[<chars>\.]* which allows trailing
dots and fixes the problem.

Bug: 2337042
Change-Id: I310512531787e0f742988f5d815ad944fd39e059